Creates recibo-actions-bar directive, recibo-compra-actions-bar component and recibo-venta-actions-bar component to reuse their functionality in recibos-venta, recibos-compra and ver-recibo components.
