+++ src/app/components/ver-recibo/ver-recibo.component.html
@@ -9,9 +9,21 @@
         <fa-icon [icon]="['fas', 'file-invoice-dollar']"></fa-icon>
         Recibo de {{ recibo.idCliente ? 'Venta' : 'Compra' }}
       </h1>
-      <button class="btn btn-primary" (click)="downloadReciboPdf()" placement="left" ngbPopover="Descargar Archivo" triggers="mouseenter:mouseleave" *ngIf="recibo.idCliente">
-        <fa-icon [icon]="['fas', 'file-download']"></fa-icon>
-      </button>
+      <div class="text-nowrap">
+        <app-recibo-compra-actions-bar *ngIf="!recibo.idCliente"
+          [hiddenButtons]="['show']"
+          [recibo]="recibo"
+          (afterDelete)="volverAlListado()">
+        </app-recibo-compra-actions-bar>
+        <app-recibo-venta-actions-bar *ngIf="recibo.idCliente" class="me-2"
+          [hiddenButtons]="['show']"
+          [recibo]="recibo"
+          (afterDelete)="volverAlListado()">
+        </app-recibo-venta-actions-bar>
+        <button class="btn btn-primary" (click)="downloadReciboPdf()" placement="left" ngbPopover="Descargar Archivo" triggers="mouseenter:mouseleave" *ngIf="recibo.idCliente">
+          <fa-icon [icon]="['fas', 'file-download']"></fa-icon>
+        </button>
+      </div>
     </div>
     <table class="table table-bordered table-sm">
       <caption class="sr-only">Datos del recibo</caption>
