+++ src/app/feature-modules/recibos-compra/components/recibos-compra/recibos-compra.component.html
@@ -98,23 +98,7 @@
       </div>
     </ng-template>
     <ng-template #actions let-item>
-      <ul class="list-inline mb-0 text-center d-inline-block">
-        <li class="list-inline-item">
-          <button class="btn btn-primary fake-cursor" ngbPopover="Ver Detalle" placement="left" triggers="mouseenter:mouseleave" (click)="verRecibo(item)">
-            <fa-icon [icon]="['far', 'eye']"></fa-icon>
-          </button>
-        </li>
-        <li class="list-inline-item" *ngIf="hasRoleToCrearNota">
-          <button class="btn btn-primary fake-cursor" ngbPopover="Crear Nota de Débito" placement="left" triggers="mouseenter:mouseleave" (click)="crearNotaDeDebitoRecibo(item)">
-            <fa-icon [icon]="['fas', 'balance-scale-right']"></fa-icon>
-          </button>
-        </li>
-        <li class="list-inline-item" *ngIf="hasRoleToDelete">
-          <button class="btn btn-primary fake-cursor" ngbPopover="Eliminar" placement="left" triggers="mouseenter:mouseleave" (click)="eliminarRecibo(item)">
-            <fa-icon [icon]="['fas', 'trash']"></fa-icon>
-          </button>
-        </li>
-      </ul>
+      <app-recibo-compra-actions-bar [recibo]="item" (afterDelete)="afterReciboDelete()"></app-recibo-compra-actions-bar>
     </ng-template>
     <app-lista [size]="size" [items]="items" [infoTemplate]="info" [actionsTemplate]="actions"
                [page]="page" [totalPages]="totalPages" [totalElements]="totalElements" (pageChange)="loadPage($event)"></app-lista>
