+++ src/app/feature-modules/recibos-venta/recibos-venta.module.ts
@@ -2,13 +2,13 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { RecibosVentaRoutingModule } from './recibos-venta-routing.module';
-import { RecibosVentaComponent } from './componentes/recibos-venta/recibos-venta.component';
+import { RecibosVentaComponent } from './components/recibos-venta/recibos-venta.component';
 import {ShareModule} from '../../modules/share.module';
 
 
 @NgModule({
   declarations: [
-    RecibosVentaComponent
+    RecibosVentaComponent,
   ],
   imports: [
     CommonModule,
